Bot refactored also changed conception to a more Spring similar way

+++ yearPercentages/src/main/resources/db/migration/V2__create_users_schema.sql
@@ -1,6 +1,6 @@
 create table if not exists users
 (
-    chat_id    serial primary key,
+    chat_id    BIGINT primary key,
     user_name  varchar(20) not null,
     is_started boolean
 );
